Added files
The .py file contains the action executable code. Requires the Mutagen module.

The Library.xml file is my original exported library as created by iTunes. I simply copied that file and saved it as .plist instead of .xml, which is where the Library.plist file comes from.

The Library.plist file is what is read by the program, it provides an index of all the current information associated with each MP3 file.